First LeetCode
A bit hard at first, shorter than you might think. Took me like half an hour to see that if you can finish the loop from one point (becouse of the condition that there is only 1 right answer) without making it below 0 it is a right point and a right answer.