121. Best Time to Buy and Sell Stock
Pretty easy one, you set lowest and highest as two pointers and you move highest to right in every step of loop, you move lowest if its lower than highest, if its not you see if thier profit is greater than previous maxProfit, if it is you set it as a new maxProfit. While lasts till the end of array, and if you cant make profit default maxProfit just return 0.